Refactor savings calculation
