+++ src/plays/savings-calculator/components/SavingOptions/index.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import { calculateSavings } from "../../utils/calculateSavings";
 import "./savingOptions.scss";
 
 function SavingOptions(props) {
@@ -13,19 +14,13 @@ function SavingOptions(props) {
   const handleSubmit = (event) => {
     event.preventDefault();
 
-    let total = parseFloat(startingBalance);
-    let monthlyInterest = parseFloat(annualInterest) / 12;
-    let monthlyInterestPaid = 0;
-    let periodInMonths =
-      periodChoice === "Years" ? parseFloat(period) * 12 : parseFloat(period);
-    let monthlyContributionAmount = parseFloat(monthlyContribution);
-
-    for (let i = 0; i < periodInMonths; i++) {
-      monthlyInterestPaid = total * (monthlyInterest / 100);
-      total = total + monthlyContributionAmount + monthlyInterestPaid;
-    }
-
-    total = parseInt(total);
+    const total = calculateSavings({
+      startingBalance,
+      monthlyContribution,
+      period,
+      periodChoice,
+      annualInterest,
+    });
 
     setTotal(total);
   };
